% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WidestBounds.R
\name{WidestBounds}
\alias{WidestBounds}
\title{Widest identification bounds}
\usage{
WidestBounds(
  database,
  X,
  Y,
  Z,
  Y0,
  Y1,
  v = NULL,
  meanmodel = "lm",
  linkprob = "logit"
)
}
\arguments{
\item{database}{data frame containing the data.}

\item{X}{vector with the number of the columns in database that are used as covariates X. Must be of length 1 or 2.}

\item{Y}{number of the column in database indicating the response variable.}

\item{Z}{number of the column in database that indicates the binary variable Z in data.}

\item{Y0}{The minimum possible value in the range of Y.}

\item{Y1}{The maximum possible value in the range of Y.}

\item{v}{dataframe indicating the values to make the prediction of Y. The number of columns must be the same as the length of X.}

\item{meanmodel}{model used to estimate the regression model \eqn{E(Y\mid X,Z=1)}. By default is a linear model (lm).}

\item{linkprob}{model used to estimate the probability \eqn{P(Z=1\mid X)}.}
}
\value{
\code{data.bounds} A data frame that contains the observed values of X and Y. Additionally it contains predicted values of Y, the probability of observing Y, and the lower and upper bounds.

\code{plot.bounds} A plot with the widest identification bounds.

\code{widthBound} If v is provided, the width of the bound is shown for each row of v.
}
\description{
This function computes the widest identification bounds of the conditional expectation \eqn{E(Y\mid X)} when some values of \eqn{Y} are not observed.
We use the partial identifiability approach introduced by Manski (1989). Let \eqn{Z} be a binary random variable such that \eqn{Z=1} if \eqn{Y} is observed, and 0 if not.
}
\details{
By the law of total probability, we know that
\deqn{E(Y\mid X)=E(Y\mid X,Z=0)P(Z=0\mid X)+E(Y\mid X,Z=1)P(Z=1\mid X)}
However, \eqn{E(Y\mid X,Z=0)} is not observed because it depends on the non observed values of \eqn{Y}.

If \eqn{Y} is bounded by \eqn{Y0} and \eqn{Y1} (i.e., \eqn{Y0\leq Y\leq Y1}), then

\deqn{Y0 P(Z=0\mid X)+E(Y\mid X,Z=1)P(Z=1\mid X)\leq E(Y\mid X)\leq Y1 P(Z=0\mid X)+E(Y\mid X,Z=1)P(Z=1\mid X)}

In this function, the identification bounds for specific values for Y0 and Y1 are computed.
}
\examples{
data("DataSelection")

WidestBounds(DataSelection,X=c(6,5),Y=2,Z=7,Y0=1,Y1=7,meanmodel = "lm",linkprob = "probit", v=data.frame(c(650,700),c(650,600)))

WidestBounds(DataSelection,X=6,Y=2,Z=7,Y0=1,Y1=7,meanmodel = "lm",linkprob = "probit", v=data.frame(c(650,700)))
}
\references{
Manski, C. (1989). Anatomy of the selection problem. The Journal of Human Resources, 24(3), 343–360.
}
